+++ src/main/java/com/heroku/spacey/dao/ProductDetailDao.java
@@ -3,11 +3,11 @@ package com.heroku.spacey.dao;
 import com.heroku.spacey.entity.ProductDetail;
 
 public interface ProductDetailDao {
-    ProductDetail getById(int id);
+    ProductDetail getById(Long id);
 
-    int insert(ProductDetail productDetail);
+    Long insert(ProductDetail productDetail);
 
     void update(ProductDetail productDetail);
 
-    void delete(int id);
+    void delete(Long id);
 }
