+++ src/main/java/com/heroku/spacey/services/CategoryService.java
@@ -3,11 +3,11 @@ package com.heroku.spacey.services;
 import com.heroku.spacey.dto.category.CategoryDto;
 
 public interface CategoryService {
-    CategoryDto getById(int id);
+    CategoryDto getById(Long id);
 
     void addCategory(CategoryDto categoryDto);
 
-    void updateCategory(CategoryDto categoryDto, int id);
+    void updateCategory(CategoryDto categoryDto, Long id);
 
-    void deleteCategory(int id);
+    void deleteCategory(Long id);
 }
